% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModels.r
\name{add_gelman}
\alias{add_gelman}
\title{@title add_gelman
 @description Calculate gelman diagnostics to assess convergence}
\usage{
add_gelman(read_in, rank.name)
}
\description{
@title add_gelman
 @description Calculate gelman diagnostics to assess convergence
}
