% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{assign_fg_categories}
\alias{assign_fg_categories}
\title{@title 			assign_fg_categories
 @description Assign functional group categories based on group name}
\usage{
assign_fg_categories(vector)
}
\description{
@title 			assign_fg_categories
 @description Assign functional group categories based on group name
}
