% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{assign_fg_kingdoms}
\alias{assign_fg_kingdoms}
\title{@title 			assign_fg_kingdoms
 @description Assign kingdoms based on functional group categories}
\usage{
assign_fg_kingdoms(vector)
}
\description{
@title 			assign_fg_kingdoms
 @description Assign kingdoms based on functional group categories
}
