% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{assign_fg_sources}
\alias{assign_fg_sources}
\title{@title 			assign_fg_sources
 @description Assign sources based on functional group development method}
\usage{
assign_fg_sources(vector)
}
\description{
@title 			assign_fg_sources
 @description Assign sources based on functional group development method
}
