% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology.r
\name{assign_pheno_category}
\alias{assign_pheno_category}
\title{@title 			assign_pheno_category
 @description Assign phenological category based on date and siteID dataframe}
\usage{
assign_pheno_category(site_date_df, pheno_df = pheno_categories_long)
}
\description{
@title 			assign_pheno_category
 @description Assign phenological category based on date and siteID dataframe
}
