% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology_optimized.r
\name{assign_pheno_category_optimized}
\alias{assign_pheno_category_optimized}
\title{@title 			assign_pheno_category_optimized
 @description Optimized version of assign_pheno_category for large datasets}
\usage{
assign_pheno_category_optimized(site_date_df, pheno_df = pheno_categories_long)
}
\description{
@title 			assign_pheno_category_optimized
 @description Optimized version of assign_pheno_category for large datasets
}
