% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology.r
\name{assign_pheno_date}
\alias{assign_pheno_date}
\title{@title 			assign_pheno_date
 @description Assign phenological category based only on date  dataframe}
\usage{
assign_pheno_date(date, pheno_df = pheno_categories_long)
}
\description{
@title 			assign_pheno_date
 @description Assign phenological category based only on date  dataframe
}
