% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology.r
\name{assign_pheno_date_batch}
\alias{assign_pheno_date_batch}
\title{@title 			assign_pheno_date_batch
 @description Batch processing version for very large datasets}
\usage{
assign_pheno_date_batch(
  dates,
  pheno_df = pheno_categories_long,
  batch_size = 1000
)
}
\description{
@title 			assign_pheno_date_batch
 @description Batch processing version for very large datasets
}
