% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology.r
\name{assign_pheno_date_vectorized}
\alias{assign_pheno_date_vectorized}
\title{@title 			assign_pheno_date_vectorized
 @description Optimized vectorized version of assign_pheno_date for multiple dates}
\usage{
assign_pheno_date_vectorized(dates, pheno_df = pheno_categories_long)
}
\description{
@title 			assign_pheno_date_vectorized
 @description Optimized vectorized version of assign_pheno_date for multiple dates
}
