% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology_optimized.r
\name{assign_pheno_date_vectorized_optimized}
\alias{assign_pheno_date_vectorized_optimized}
\title{@title 			assign_pheno_date_vectorized_optimized
 @description Highly optimized vectorized version of assign_pheno_date for large datasets}
\usage{
assign_pheno_date_vectorized_optimized(dates, pheno_df = pheno_categories_long)
}
\description{
@title 			assign_pheno_date_vectorized_optimized
 @description Highly optimized vectorized version of assign_pheno_date for large datasets
}
