% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology.r
\name{assign_pheno_site_date}
\alias{assign_pheno_site_date}
\title{@title 			assign_pheno_site_date
 @description Assign phenological category based on date and siteID}
\usage{
assign_pheno_site_date(
  site = "HARV",
  date_ymd = "2015-01-01",
  pheno_df = pheno_categories_long
)
}
\description{
@title 			assign_pheno_site_date
 @description Assign phenological category based on date and siteID
}
