% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPhenology_optimized.r
\name{assign_pheno_site_date_optimized}
\alias{assign_pheno_site_date_optimized}
\title{@title 			assign_pheno_site_date_optimized
 @description Optimized version of assign_pheno_site_date for large datasets}
\usage{
assign_pheno_site_date_optimized(
  site = "HARV",
  date_ymd = "2015-01-01",
  pheno_df = pheno_categories_long
)
}
\description{
@title 			assign_pheno_site_date_optimized
 @description Optimized version of assign_pheno_site_date for large datasets
}
