% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBetaRegModels.r
\name{calculate_plot_summary_from_samples}
\alias{calculate_plot_summary_from_samples}
\title{@title calculate_plot_summary_from_samples
 @description Calculate plot summary statistics from samples2 data
 @param samples2 MCMC samples containing plot estimates
 @return List with statistics and quantiles data frames}
\usage{
calculate_plot_summary_from_samples(samples2)
}
\description{
@title calculate_plot_summary_from_samples
 @description Calculate plot summary statistics from samples2 data
 @param samples2 MCMC samples containing plot estimates
 @return List with statistics and quantiles data frames
}
