% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{check_continue}
\alias{check_continue}
\title{check_continue}
\usage{
check_continue(run1, min_eff_size = 50)
}
\description{
Determine whether MCMC should continue running, based on the minimum effective sample size
}
