% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{combine_chains_simple}
\alias{combine_chains_simple}
\title{combine_chains_simple}
\usage{
combine_chains_simple(
  chain_paths,
  save = FALSE,
  cut_size1 = NULL,
  cut_size2 = NULL
)
}
\description{
Combine MCMC chains using paths, shortening each chain due to RAM constraints
}
