% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast.r
\docType{data}
\name{covariate_cache}
\alias{covariate_cache}
\title{Cached site covariate generation}
\format{
An object of class \code{environment} of length 0.
}
\usage{
covariate_cache
}
\description{
Cached site covariate generation
}
\keyword{datasets}
