% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast_CLR.r
\name{create_covariate_samples_clr}
\alias{create_covariate_samples_clr}
\title{create_covariate_samples_clr}
\usage{
create_covariate_samples_clr(model.inputs, plotID, siteID, Nmc_large, Nmc)
}
\description{
Create covariate samples for CLR model forecasting
}
