% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{crib_fun}
\alias{crib_fun}
\title{crib_fun
stolen from colin's NEFI_microbe repo
converts a vector of [0,1] values to (0,1) a la Cribari-Neto & Zeileis 2010}
\usage{
crib_fun(x, N = NA)
}
\arguments{
\item{x}{a vector of values on the interval [0,1]}

\item{N}{alternative sample size. This is useful when tranforming a matrix in the dirchlet case, rather than just a vector as in the beta case.}
}
\value{
a vector of values on the interval (0,1)
}
\description{
crib_fun
stolen from colin's NEFI_microbe repo
converts a vector of [0,1] values to (0,1) a la Cribari-Neto & Zeileis 2010
}
