% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{extract_bracketed_vals}
\alias{extract_bracketed_vals}
\title{extract_bracketed_vals}
\usage{
extract_bracketed_vals(input_df, varname1 = "beta_num", varname2 = NULL)
}
\description{
# extract MCMC summary by rowname from summary matrix
}
