% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineChains.r
\name{extract_metadata_from_paths}
\alias{extract_metadata_from_paths}
\title{extract_metadata_from_paths}
\usage{
extract_metadata_from_paths(chain_paths)
}
\arguments{
\item{chain_paths}{Paths to chain files}
}
\value{
Basic metadata object
}
\description{
Extract basic metadata information from file paths when no metadata is available
}
