% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{extract_summary_row}
\alias{extract_summary_row}
\title{extract_summary_row}
\usage{
extract_summary_row(input_df, var = "sigma")
}
\description{
extract MCMC summary by rowname from summary matrix
}
