% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{fast.summary.mcmc}
\alias{fast.summary.mcmc}
\title{fast.summary.mcmc}
\usage{
fast.summary.mcmc(object, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\description{
Duplicate of coda::summary.mcmc(), except using datatable for speed
}
