% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast_CLR.r
\name{fcast_clr}
\alias{fcast_clr}
\title{fcast_clr}
\usage{
fcast_clr(
  plotID,
  model.inputs,
  param_samples,
  truth.plot.long,
  plot_summary,
  Nmc = 1000,
  drop_other = T,
  predict_site_effects = NULL,
  rank.name = NULL,
  model_id,
  ...
)
}
\description{
Forecast functional groups at NEON plots and sites using CLR model output
}
