% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast_dirichlet.r
\name{fcast_dirichlet}
\alias{fcast_dirichlet}
\title{fcast_dirichlet}
\usage{
fcast_dirichlet(
  plotID,
  model.inputs,
  param_samples,
  truth.plot.long,
  plot_summary,
  Nmc = 1000,
  drop_other = T,
  predict_site_effects = NULL,
  rank.name = NULL,
  model_id,
  ...
)
}
\description{
Forecast taxonomic groups at NEON plots and sites using Dirichlet model output
}
