% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast.r
\name{fcast_logit_beta}
\alias{fcast_logit_beta}
\title{fg_fcast_beta}
\usage{
fcast_logit_beta(
  plotID,
  model.inputs,
  param_samples,
  truth.plot.long,
  plot_summary = NULL,
  Nmc = 1000,
  predict_site_effects = NULL,
  rank.name = NULL,
  model_id,
  metadata = NULL,
  col_idx = NULL,
  ...
)
}
\description{
Forecast functional groups at NEON plots and sites, using beta regression output

Unified forecasting function for observed and unobserved sites
}
