% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastFunctional_betareg.r
\name{fg_fcast_beta}
\alias{fg_fcast_beta}
\title{fg_fcast_beta}
\usage{
fg_fcast_beta(
  plotID,
  model.inputs,
  param_samples,
  truth.plot.long,
  plot_summary,
  Nmc = 1000,
  drop_other = T,
  ...
)
}
\description{
Forecast functional groups at NEON plots and sites, using beta regression output
}
