% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{filter_date_site}
\alias{filter_date_site}
\title{filter_date_site}
\usage{
filter_date_site(
  input_df,
  keep_sites,
  keep_plots,
  min.date,
  max.date,
  max.predictor.date = NULL,
  ...
)
}
\value{
a vector of values on the interval (0,1)
}
\description{
return data matrix filtered by date/plot/site
}
