% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{fixDate}
\alias{fixDate}
\title{@title 			fixDate
 @description Take a vector of dates in dateID format ("201401") and return as date with first of month}
\usage{
fixDate(datesToFix)
}
\description{
@title 			fixDate
 @description Take a vector of dates in dateID format ("201401") and return as date with first of month
}
