% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsFunctions.r
\name{get_sin_cos}
\alias{get_sin_cos}
\title{@title 			get_sin_cos
 @description return sin and cosine components from month or day}
\usage{
get_sin_cos(input_dates)
}
\arguments{
\item{input_dates}{a vector of dates in the format "20210130" or "202101"}
}
\value{
a vector of values on the interval (0,1)
}
\description{
@title 			get_sin_cos
 @description return sin and cosine components from month or day
}
\examples{
get_sin_cos("20210130")
get_sin_cos("202101")
}
