% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineChains.r
\name{infer_legacy_covariate_from_paths}
\alias{infer_legacy_covariate_from_paths}
\title{infer_legacy_covariate_from_paths}
\usage{
infer_legacy_covariate_from_paths(chain_paths)
}
\arguments{
\item{chain_paths}{Paths to chain files}
}
\value{
Boolean indicating if legacy covariate is used
}
\description{
Infer legacy covariate flag from file paths
}
