% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineChains.r
\name{infer_model_name_from_paths}
\alias{infer_model_name_from_paths}
\title{infer_model_name_from_paths}
\usage{
infer_model_name_from_paths(chain_paths)
}
\arguments{
\item{chain_paths}{Paths to chain files}
}
\value{
Inferred model name
}
\description{
Infer model name from file paths
}
