% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{interval_transform}
\alias{interval_transform}
\title{interval_transform
Take a vector of values and put them in the (0,1) interval}
\usage{
interval_transform(x, C = ncol(x), N = nrow(x))
}
\description{
interval_transform
Take a vector of values and put them in the (0,1) interval
}
