% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plot_estimates.r
\name{load_plot_estimates_for_hindcast}
\alias{load_plot_estimates_for_hindcast}
\title{Load plot estimates for hindcast initialization}
\usage{
load_plot_estimates_for_hindcast(model_type = "env_cycl")
}
\arguments{
\item{model_type}{Character string: "env_cycl" or "cycl_only"}
}
\value{
data.table with plot estimates for hindcast initialization
}
\description{
Load plot estimates for hindcast initialization
}
