% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plot_estimates.r
\name{load_plot_estimates_for_scoring}
\alias{load_plot_estimates_for_scoring}
\title{Load plot estimates for scoring metrics}
\usage{
load_plot_estimates_for_scoring(model_type = "env_cycl")
}
\arguments{
\item{model_type}{Character string: "env_cycl" or "cycl_only"}
}
\value{
data.table with plot estimates for scoring
}
\description{
Load plot estimates for scoring metrics
}
