% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plot_estimates.r
\name{load_plot_estimates_for_species}
\alias{load_plot_estimates_for_species}
\title{Load plot estimates for a specific species}
\usage{
load_plot_estimates_for_species(species_name, model_type = "env_cycl")
}
\arguments{
\item{species_name}{Character string: species name to filter by}

\item{model_type}{Character string: "env_cycl" or "cycl_only"}
}
\value{
data.table with plot estimates for the specified species
}
\description{
Load plot estimates for a specific species
}
