% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleMod_shannon}
\alias{nimbleMod_shannon}
\title{nimbleMod_shannon}
\format{
An object of class \code{list} of length 12.
}
\usage{
nimbleMod_shannon
}
\description{
nimbleMod_shannon
}
\keyword{datasets}
