% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_cycl_only.r
\docType{data}
\name{nimbleMod_shannon_cycl_only}
\alias{nimbleMod_shannon_cycl_only}
\title{NIMBLE model for Shannon diversity with cyclic covariates only}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleMod_shannon_cycl_only
}
\description{
A NIMBLE model code object for fitting Shannon diversity data using only cyclic covariates
}
\keyword{datasets}
