% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleModFunctional_trunc}
\alias{nimbleModFunctional_trunc}
\title{NIMBLE model for functional groups with truncated Normal distribution}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleModFunctional_trunc
}
\description{
A NIMBLE model code object for fitting functional group data with truncated Normal distribution to prevent negative values
}
\keyword{datasets}
