% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleModTaxa}
\alias{nimbleModTaxa}
\title{NIMBLE model for taxonomic data}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleModTaxa
}
\description{
A NIMBLE model code object for fitting taxonomic abundance data
}
\keyword{datasets}
