% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_cycl_only.r
\docType{data}
\name{nimbleModTaxa_cycl_only}
\alias{nimbleModTaxa_cycl_only}
\title{NIMBLE model for taxonomic data with cyclic covariates only}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleModTaxa_cycl_only
}
\description{
A NIMBLE model code object for fitting taxonomic abundance data using only cyclic covariates
}
\keyword{datasets}
