% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCLRModels.r
\name{parse_clr_model_id}
\alias{parse_clr_model_id}
\title{parse_clr_model_id}
\usage{
parse_clr_model_id(model_id)
}
\arguments{
\item{model_id}{The model ID string}
}
\description{
Parse CLR model ID for summary functions
}
