% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsFunctions.r
\name{predictive_loss}
\alias{predictive_loss}
\title{predictive_loss}
\usage{
predictive_loss(observed, predicted, predicted_sd)
}
\description{
Calculate predictive loss decomposed into predictive variance and residual variance
}
