% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDiversityData.r
\name{prepDivData}
\alias{prepDivData}
\title{prepDivData}
\usage{
prepDivData(
  rank.df,
  min.date = "20130601",
  max.date = "20170101",
  predictor_data = readRDS("./data/clean/all_predictor_data.rds"),
  min.prev = 5,
  dom_soil_horizons = readRDS("./data/clean/dominantHorizonsSite.rds"),
  full_timeseries = F
)
}
\description{
prepDivData
}
