% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFunctionalData.r
\name{prepFunctionalData}
\alias{prepFunctionalData}
\title{prepFunctionalData}
\usage{
prepFunctionalData(
  rank.df,
  min.date = "20130601",
  max.date = "20170101",
  predictor_data = NULL,
  min.prev = 5,
  dom_soil_horizons = NULL,
  full_timeseries = F
)
}
\description{
prepFunctionalData
}
