% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTaxonomicData.r
\name{prepTaxonomicData}
\alias{prepTaxonomicData}
\title{prepTaxonomicData}
\usage{
prepTaxonomicData(
  rank.df,
  min.date = "20130601",
  max.date = "20170101",
  predictor_data = NULL,
  min.prev = 3,
  dom_soil_horizons = NULL,
  full_timeseries = FALSE,
  keep_vec = NULL
)
}
\description{
prepTaxonomicData
}
