% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{rbind.df.list}
\alias{rbind.df.list}
\title{@title rbind.df.list
 @description rbind named dfs}
\usage{
\method{rbind}{df.list}(pl.out)
}
\description{
@title rbind.df.list
 @description rbind named dfs
}
