% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCMC_bychain_CLR.r
\name{run_MCMC_bychain_CLR}
\alias{run_MCMC_bychain_CLR}
\title{run_spatial_model}
\usage{
run_MCMC_bychain_CLR(
  rank.name = "phylum_bac",
  s = "acidobacteriota",
  temporalDriverUncertainty = TRUE,
  spatialDriverUncertainty = TRUE,
  scenario = NULL,
  min.date = "20151101",
  max.date = "20180101",
  model_name = "spatial",
  model_id = NULL,
  test = F,
  chain_no = 1,
  burnin = 500,
  thin = 1,
  thin2 = 20,
  init_iter = 10000,
  iter_per_chunk = 5000,
  min_eff_size_perchain = 5,
  max_loops = 100,
  max_save_size = 40000,
  ...
)
}
\description{
run_spatial_model
}
