% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_taxa.r
\name{run_MCMC_dirichlet}
\alias{run_MCMC_dirichlet}
\title{run_MCMC_dirichlet}
\usage{
run_MCMC_dirichlet(
  k = 1,
  iter = 1000,
  init_iter = 1e+05,
  iter_per_chunk = 10000,
  burnin = 500,
  thin = 1,
  test = F,
  temporalDriverUncertainty = TRUE,
  spatialDriverUncertainty = TRUE,
  scenario = NULL,
  s = "acidobacteriota",
  model_name = "all_covariates",
  chain_no = 1,
  max.date = "20180101",
  min.date = "20151101",
  ...
)
}
\description{
run_MCMC_dirichlet

run_MCMC_taxa for entire rank (more than one taxon)
}
