% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_diversity.r
\name{run_MCMC_diversity}
\alias{run_MCMC_diversity}
\title{run_MCMC_diversity}
\usage{
run_MCMC_diversity(
  group = "ITS",
  iter = 1000,
  burnin = 500,
  thin = 1,
  test = F,
  n.chains = 3,
  temporalDriverUncertainty = TRUE,
  spatialDriverUncertainty = TRUE,
  scenario = NULL,
  min.date = "20160101",
  max.date = "20180101",
  model_name = "cycl_only",
  time_period = "calibration"
)
}
\description{
run_MCMC_diversity
}
