% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_functional_groups.r
\name{run_MCMC_functional}
\alias{run_MCMC_functional}
\title{run_MCMC_functional}
\usage{
run_MCMC_functional(
  k = 17,
  iter = 1000,
  burnin = 500,
  thin = 1,
  test = F,
  temporalDriverUncertainty = TRUE,
  spatialDriverUncertainty = TRUE,
  scenario = NULL,
  model_name = "cycl_only",
  time_period = "calibration",
  max.date = "20200101",
  min.date = "20160101",
  ...
)
}
\description{
run_MCMC_functional
}
