% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_single_taxon.r
\name{run_MCMC_single_taxon}
\alias{run_MCMC_single_taxon}
\title{run_MCMC_single_taxon}
\usage{
run_MCMC_single_taxon(
  k = 1,
  iter = 1000,
  burnin = 500,
  thin = 1,
  test = F,
  temporalDriverUncertainty = TRUE,
  spatialDriverUncertainty = TRUE,
  scenario = NULL,
  s = "acidobacteriota",
  min.date = "20151101",
  max.date = "20180101",
  model_name = "cycl_only",
  ...
)
}
\description{
run_MCMC_single_taxon
}
