% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_single_taxon_bychain.r
\name{run_MCMC_single_taxon_bychain}
\alias{run_MCMC_single_taxon_bychain}
\title{run_MCMC_single_taxon_bychain}
\usage{
run_MCMC_single_taxon_bychain(
  k = 1,
  iter = 1000,
  burnin = 500,
  thin = 1,
  test = F,
  chain_no = 1,
  temporalDriverUncertainty = TRUE,
  spatialDriverUncertainty = TRUE,
  scenario = NULL,
  s = "acidobacteriota",
  min.date = "20151101",
  max.date = "20180101",
  model_name = "cycl_only",
  iter_per_chunk = 5000,
  init_iter = 10000,
  ...
)
}
\description{
run_MCMC_single_taxon_bychain
}
