% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_single_taxon_bychain.r
\docType{data}
\name{s}
\alias{s}
\title{run_MCMC_single_taxon_bychain}
\format{
An object of class \code{character} of length 1.
}
\usage{
s
}
\description{
run_MCMC_single_taxon_bychain
}
\keyword{datasets}
