% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModels.r
\name{samples_exists}
\alias{samples_exists}
\title{@title samples_exists
 @description Check if samples file (condensing all 4 chains) already exists}
\usage{
samples_exists(file_path)
}
\description{
@title samples_exists
 @description Check if samples file (condensing all 4 chains) already exists
}
