% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsFunctions.r
\name{sin_cos_to_seasonality}
\alias{sin_cos_to_seasonality}
\title{sin_cos_to_seasonality}
\usage{
sin_cos_to_seasonality(sin, cos)
}
\description{
Convert sin and cos effect sizes to seasonal amplitude parameter
}
