% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBetaRegModels.r
\name{summarize_beta_model}
\alias{summarize_beta_model}
\title{@title summarize_beta_regression
 @description Summarize NIMBLE state-space beta regression models for microbial taxa and functional groups
 Assumes input RDS files contain a list of:
 MCMC samples, parameter summaries, latent state summaries, and model-fitting metadata}
\usage{
summarize_beta_model(
  file_path,
  save_summary = NULL,
  overwrite = NULL,
  drop_other = TRUE
)
}
\arguments{
\item{overwrite}{want to save new summary files even if there's an existing, recent summary file}
}
\description{
@title summarize_beta_regression
 @description Summarize NIMBLE state-space beta regression models for microbial taxa and functional groups
 Assumes input RDS files contain a list of:
 MCMC samples, parameter summaries, latent state summaries, and model-fitting metadata
}
