% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCLRModels.r
\name{summarize_clr_model}
\alias{summarize_clr_model}
\title{@title summarize_clr_model
 @description Summarize NIMBLE state-space CLR models for microbial taxa and functional groups
 Assumes input RDS files contain a list of:
 MCMC samples, parameter summaries, latent state summaries, and model-fitting metadata}
\usage{
summarize_clr_model(
  file_path,
  save_summary = NULL,
  overwrite = NULL,
  drop_other = TRUE
)
}
\arguments{
\item{overwrite}{want to save new summary files even if there's an existing, recent summary file}
}
\description{
@title summarize_clr_model
 @description Summarize NIMBLE state-space CLR models for microbial taxa and functional groups
 Assumes input RDS files contain a list of:
 MCMC samples, parameter summaries, latent state summaries, and model-fitting metadata
}
