% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModels.r
\name{summarize_fg_beta_model}
\alias{summarize_fg_beta_model}
\title{@title summarize_fg_beta_model
 @description Summarize beta regression models for functional groups}
\usage{
summarize_fg_beta_model(file_path, save_summary = NULL, drop_other = TRUE)
}
\description{
@title summarize_fg_beta_model
 @description Summarize beta regression models for functional groups
}
