% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModels.r
\name{summarize_fg_div_model}
\alias{summarize_fg_div_model}
\title{@title summarize_fg_div_model
 @description summarize_fg_div_model}
\usage{
summarize_fg_div_model(file_path, drop_other = TRUE)
}
\description{
@title summarize_fg_div_model
 @description summarize_fg_div_model
}
