% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModels.r
\name{summarize_tax_model}
\alias{summarize_tax_model}
\title{summarize_tax_model}
\usage{
summarize_tax_model(file_path, save_summary = NULL, drop_other = TRUE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
summarize_tax_model
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
