% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeModels.r
\name{summary_exists}
\alias{summary_exists}
\title{@title summary_exists
 @description Check if summary file already exists}
\usage{
summary_exists(file_path)
}
\description{
@title summary_exists
 @description Check if summary file already exists
}
