% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsFunctions.r
\name{tag_facet}
\alias{tag_facet}
\title{tag_facet}
\usage{
tag_facet(
  p,
  open = "(",
  close = ")",
  tag_pool = letters,
  x = -Inf,
  y = Inf,
  hjust = -0.5,
  vjust = 1.5,
  fontface = 2,
  family = "",
  ...
)
}
\description{
Add letter tags to ggplot facet panels
}
