% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{toc}
\alias{toc}
\title{toc}
\usage{
toc()
}
\description{
one of two clock functions.
Place tic() at the line in the code where you want to start timing.
Place toc() at the position in the code where you want to stop timing and report.
}
