% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast.r
\name{validate_forecast_inputs}
\alias{validate_forecast_inputs}
\title{Comprehensive validation function for forecasting inputs}
\usage{
validate_forecast_inputs(param_samples, model.inputs, plotID, model_name, Nmc)
}
\arguments{
\item{param_samples}{Parameter samples matrix}

\item{model.inputs}{Model inputs list}

\item{plotID}{Plot identifier}

\item{model_name}{Model type name}

\item{Nmc}{Number of Monte Carlo samples}
}
\value{
List of validation results
}
\description{
Comprehensive validation function for forecasting inputs
}
