% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hindcast.r
\name{vectorized_forecast}
\alias{vectorized_forecast}
\title{Vectorized Monte Carlo simulation (robust, supports cloglog + legacy)}
\usage{
vectorized_forecast(
  params,
  covar,
  initial_conditions,
  timepoints,
  Nmc,
  use_cloglog = FALSE,
  legacy = NULL
)
}
\description{
Vectorized Monte Carlo simulation (robust, supports cloglog + legacy)
}
