% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsFunctions.r
\name{add_scoring_metrics}
\alias{add_scoring_metrics}
\title{add_scoring_metrics}
\usage{
add_scoring_metrics(
  observed,
  mean_predicted,
  sd_predicted,
  type = c("RMSE", "BIAS", "MAE", "CRPS", "RSQ", "RSQ.1", "RMSE.norm",
    "residual_variance", "predictive_variance", "total_PL", "CRPS_truncated")
)
}
\description{
Score models and forecasts
}
