% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{combine_chains}
\alias{combine_chains}
\title{combine_chains}
\usage{
combine_chains(chain_paths, save = FALSE, cut_size1 = NULL, cut_size2 = NULL)
}
\description{
Combine MCMC chains using paths, shortening each chain due to RAM constraints
}
