% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.r
\name{create_covariate_samples}
\alias{create_covariate_samples}
\title{create_covariate_samples}
\usage{
create_covariate_samples(
  model.inputs,
  plotID = NULL,
  siteID,
  Nmc_large,
  Nmc,
  N.beta = 8,
  prev_samples = NULL,
  ...
)
}
\description{
samples covariate values
}
