% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleMod_shannon}
\alias{nimbleMod_shannon}
\title{NIMBLE model for Shannon diversity}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleMod_shannon
}
\description{
A NIMBLE model code object for fitting Shannon diversity data
}
\keyword{datasets}
