% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_cycl_only.r
\docType{data}
\name{nimbleMod_shannon_cycl_only}
\alias{nimbleMod_shannon_cycl_only}
\title{nimbleMod_shannon_cycl_only}
\format{
An object of class \code{list} of length 10.
}
\usage{
nimbleMod_shannon_cycl_only
}
\description{
nimbleMod_shannon_cycl_only
}
\keyword{datasets}
