% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleModFunctional}
\alias{nimbleModFunctional}
\title{NIMBLE model for functional group data}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleModFunctional
}
\description{
A NIMBLE model code object for fitting functional group abundance data
}
\keyword{datasets}
