% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleModFunctional}
\alias{nimbleModFunctional}
\title{nimbleModFunctional}
\format{
An object of class \code{{} of length 12.
}
\usage{
nimbleModFunctional
}
\description{
nimbleModFunctional
}
\keyword{datasets}
