% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_cycl_only.r
\docType{data}
\name{nimbleModFunctional_cycl_only}
\alias{nimbleModFunctional_cycl_only}
\title{NIMBLE model for functional group data with cyclic covariates only}
\format{
An object of class \code{list} containing NIMBLE model code
}
\usage{
nimbleModFunctional_cycl_only
}
\description{
A NIMBLE model code object for fitting functional group abundance data using only cyclic covariates
}
\keyword{datasets}
