% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_cycl_only.r
\docType{data}
\name{nimbleModFunctional_cycl_only}
\alias{nimbleModFunctional_cycl_only}
\title{nimbleModFunctional_cycl_only}
\format{
An object of class \code{{} of length 10.
}
\usage{
nimbleModFunctional_cycl_only
}
\description{
nimbleModFunctional_cycl_only
}
\keyword{datasets}
