% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleModFunctional_trunc}
\alias{nimbleModFunctional_trunc}
\title{nimbleModFunctional_trunc}
\format{
An object of class \code{{} of length 12.
}
\usage{
nimbleModFunctional_trunc
}
\description{
Model version with truncated Normal distribution to prevent negative values
}
\keyword{datasets}
