% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels.r
\docType{data}
\name{nimbleModTaxa}
\alias{nimbleModTaxa}
\title{nimbleModTaxa}
\format{
An object of class \code{list} of length 10.
}
\usage{
nimbleModTaxa
}
\description{
nimbleModTaxa
}
\keyword{datasets}
