% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleModels_cycl_only.r
\docType{data}
\name{nimbleModTaxa_cycl_only}
\alias{nimbleModTaxa_cycl_only}
\title{nimbleModTaxa_cycl_only}
\format{
An object of class \code{{} of length 8.
}
\usage{
nimbleModTaxa_cycl_only
}
\description{
nimbleModTaxa_cycl_only
}
\keyword{datasets}
