% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastTaxonomic.r
\name{taxa_fcast}
\alias{taxa_fcast}
\title{taxa_fcast}
\usage{
taxa_fcast(
  plotID,
  model.inputs,
  param_samples,
  truth.plot.long,
  plot_summary,
  drop_other = T,
  predict_site_effects = NULL,
  Nmc = 1000,
  ...
)
}
\description{
Forecast taxa at NEON plots and sites
}
